+++ com.unity.perception/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
@@ -19,18 +20,34 @@ Fixed compilation warnings with latest com.unity.simulation.core package.
 
 ### Security
 
+## [0.3.0-preview.1] - 2020-07-31
+
+### Added
+
+Added realtime visualization capability to the perception package.
+
+Added visualizers for built-in labelers: Semantic Segmentation, 2D Bounding Boxes, Object Count, and Render Object Info.
+
 ## [0.2.0-preview.2] - 2020-07-15
+
 ### Fixed
+
 Fixed bug that prevented RGB captures to be written out to disk
 Fixed compatibility with com.unity.simulation.capture@0.0.10-preview.8
 
 ## [0.2.0-preview.1] - 2020-07-02
+
 ### Added
+
 Added CameraLabeler, an extensible base type for all forms of dataset output from a camera.
 Added LabelConfig\<T\>, a base class for mapping labels to data used by a labeler. There are two new derived types - ID label config and semantic segmentation label config.
+
 ### Changed
+
 Moved the various forms of ground truth from PerceptionCamera into various subclasses of CameraLabeler.
 Renamed SimulationManager to DatasetCapture.
 Changed Semantic Segmentation to take a SemanticSegmentationLabelConfig, which maps labels to color pixel values.
+
 ## [0.1.0] - 2020-06-24
+
 ### This is the first release of the _Perception_ package